requirements.txt
Django==4.1.7
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
mssql-django==1.2
mysqlclient==2.1.1
python-dotenv==0.21.0
python-json-logger==2.0.4
